feat: Prepare project for open source release (v6.0.0)
## Docker Setup
- Add production-ready docker-compose.yml with PostgreSQL 16, Redis 7, Nginx, PHP-FPM
- Add multi-stage Dockerfile with development, production, and testing stages
- Add docker-compose.override.yml.example for development environment
- Add Docker configuration files (nginx, php, postgres)
- Add entrypoint scripts and health checks

## Documentation
- Rewrite README.md as professional open source README
- Remove all sensitive data (domains, IPs) while keeping NileStack attribution
- Add CONTRIBUTING.md with coding standards and PR guidelines
- Add LICENSE (MIT License - Copyright NileStack)
- Add SECURITY.md with vulnerability reporting policy
- Add GitHub templates (bug report, feature request, PR template)
- Add docs/CONFIGURATION.md with comprehensive configuration guide

## Configuration Improvements
- Enhance .env.example with 400+ lines of documented configuration
- Add .env.docker for Docker-specific deployments
- Extend config/devflow.php with logging, metrics, security, frameworks settings
- Add config/services.php enhancements for git providers

## Code Quality
- Upgrade PHPStan to Level 9 with 0 errors
- Add HealthScoreMapper class for centralized health status mapping
- Add ProjectHealthScorer class (extracted from ProjectHealthService)
- Fix type safety issues across 19 files
- Add comprehensive ignore patterns for Laravel-specific patterns

## New Features
- Add Dashboard component refactoring (Stats, Quick Actions, Recent Activity, Server Health)
- Add CheckProjectHealthJob for async health checks
- Add keyboard shortcuts component
- Add toast notification component
- Add database migration for composite indexes

## Tests
- Add 50+ new test files for comprehensive coverage
- Add API authentication and rate limiting tests
- Add integration tests for backup, CI/CD, SSL, multi-project deployments
- Add Livewire component tests

+++ app/Services/StorageService.php
@@ -50,7 +50,7 @@ class StorageService
     public function getTotalStorageUsed(): array
     {
         $totalMB = Project::sum('storage_used_mb');
-        $maxGB = config('app.max_storage_gb', 100);
+        $maxGB = (int) config('app.max_storage_gb', 100);
         $maxMB = $maxGB * 1024;
 
         return [
